+++ NEWS.rst
@@ -6,7 +6,13 @@ of every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Added support for binary16 finite field.
+* Patch: Fix defines so that the library now works on Windows.
+* Major: Switch from using kodo-rlnc's mix-in layer stacks to using the
+  wrappers.
+* Major: Upgraded to kodo-rlnc 11.
+* Major: Changed c "namespace" from ``kodo_rlnc`` to ``krlnc``.
+* Major: Changed ``construct`` and ``destruct`` to ``new`` and ``delete``.
 
 1.0.0
 -----